Use VehicleTags to reflect the API naming convension.
