+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/checkout/checkoutActivity/views/CheckoutView.kt
@@ -31,7 +31,7 @@ import com.karhoo.uisdk.screen.booking.domain.address.BookingStatus
 import com.karhoo.uisdk.screen.booking.quotes.extendedcapabilities.Capability
 import com.karhoo.uisdk.service.preference.KarhooPreferenceStore
 import com.karhoo.uisdk.util.DateUtil
-import com.karhoo.uisdk.util.TagType
+import com.karhoo.uisdk.util.VehicleTags
 import com.karhoo.uisdk.util.extension.hideSoftKeyboard
 import com.karhoo.uisdk.util.extension.isGuest
 import kotlinx.android.synthetic.main.uisdk_booking_checkout_view.view.bookingCheckoutPassengerView
@@ -186,7 +186,7 @@ internal class CheckoutView @JvmOverloads constructor(context: Context,
                                              vehicle.vehicle.vehicleClass.orEmpty(),
                                              vehicle.serviceAgreements?.freeCancellation,
                                              vehicle.vehicle.vehicleTags.map {
-                                                 return@map TagType(it)
+                                                 return@map VehicleTags(it)
                                              },
                                              vehicle.fleet.description,
                                              isPrebook)
@@ -226,7 +226,7 @@ internal class CheckoutView @JvmOverloads constructor(context: Context,
         bookingRequestQuotesWidget.setCapacity(
                 luggage = vehicle.luggageCapacity,
                 people = vehicle.passengerCapacity,
-                otherCapabilities = capabilities.size
+                capabilitiesCount = capabilities.size
                                               )
 
         bookingRequestQuotesWidget.setCapabilities(capabilities)
