+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/checkout/quotes/BookingQuotesMVP.kt
@@ -6,12 +6,12 @@ import android.graphics.drawable.Drawable
 import android.text.Spannable
 import com.karhoo.sdk.api.model.ServiceCancellation
 import com.karhoo.uisdk.screen.booking.quotes.extendedcapabilities.Capability
-import com.karhoo.uisdk.util.TagType
+import com.karhoo.uisdk.util.VehicleTags
 
 interface BookingQuotesMVP {
 
     interface View {
-        fun setCapacity(luggage: Int, people: Int, otherCapabilities: Int)
+        fun setCapacity(luggage: Int, people: Int, capabilitiesCount: Int)
         fun setCancellationText(text: String)
         fun setCategoryText(text: String)
         fun showCancellationText(show: Boolean)
@@ -22,6 +22,6 @@ interface BookingQuotesMVP {
     interface Presenter {
         fun checkCancellationSLAMinutes(context: Context, serviceCancellation: ServiceCancellation?, isPrebook: Boolean)
         fun capitalizeCategory(category: String)
-        fun createTagsString(tags: List<TagType>, resources: Resources, shortVersion: Boolean): Spannable
+        fun createTagsString(tags: List<VehicleTags>, resources: Resources, shortVersion: Boolean): Spannable
     }
 }
